single-service check using k8s client fetching host
