+++ go.mod
@@ -5,5 +5,6 @@ go 1.13
 require (
 	github.com/pkg/errors v0.8.1
 	github.com/spf13/cobra v0.0.5
+	k8s.io/apimachinery v0.0.0-20190612205821-1799e75a0719
 	k8s.io/client-go v0.0.0-20190620085101-78d2af792bab
 )
